chore(*): Moved dev folder to root
