+++ .gitignore
@@ -12,6 +12,7 @@ node_modules
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+firebase-debug.log*
 
 # Editor directories and files
 .idea
