+++ dev/functions/src/index.ts
@@ -1,7 +1,7 @@
 import * as functions from "firebase-functions";
 import * as admin from "firebase-admin";
-import { getAllUsers as getAllUsersHandler } from "../../../cloud-functions";
-import { REGION } from "../../../constants/firebase";
+import { getAllUsers as getAllUsersHandler } from "../../../src/cloud-functions";
+import { REGION } from "../../../src/constants/firebase";
 
 // https://firebase.google.com/docs/functions/typescript
