+++ package.json
@@ -14,7 +14,7 @@
     "!src/dev"
   ],
   "scripts": {
-    "serve": "vue-cli-service serve src/dev/main.ts",
+    "serve": "vue-cli-service serve dev/main.ts",
     "build": "cross-env NODE_ENV=production rollup --config build/rollup.config.js",
     "build:ssr": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format cjs",
     "build:es": "cross-env NODE_ENV=production rollup --config build/rollup.config.js --format es",
